Fix build on Windows broken after adding support for Mac OS X.

+++ jorney/zmq/basics/cpp/main/include/basics/Thread.h
@@ -1,6 +1,6 @@
 #pragma once
 
 namespace basics {
-	void sleep(size_t seconds);
-	void msleep(size_t milliSeconds);
+	void sleep(unsigned int seconds);
+	void msleep(unsigned int milliSeconds);
 }
